first updates

+++ torch_brain/utils/binning.py
@@ -11,7 +11,7 @@ def bin_spikes(
     max_spikes: Optional[int] = None,
     right: bool = True,
     eps: float = 1e-3,
-    dtype: np.dtype = np.float32,
+    dtype: np.dtype = np.dtype,
 ) -> np.ndarray:
     r"""Bins spikes into time bins of size `bin_size`. If the total time spanned by
     the spikes is not a multiple of `bin_size`, the spikes are truncated to the nearest
